+++ packages/clickhouse/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @hypequery/core Changelog
+# @hypequery/clickhouse Changelog
 
 # [0.2.0](https://github.com/lureilly1/hypequery/compare/v0.1.0...v0.2.0) (2025-03-17)